Rename testing to integration
