+++ .github/workflows/tests.yaml
@@ -30,4 +30,4 @@ jobs:
         with:
           go-version: "1.20"
       - name: Test
-        run: go test -race -v ./testing
+        run: go test -race -v ./integration
